+++ app.js
@@ -5,7 +5,7 @@ const logger = require("morgan");
 
 const indexRouter = require("./routes/index");
 const usersRouter = require("./routes/users");
-const apiRouter = express.Router();
+const apiRouter = require("./routes/api");
 
 const app = express();
 
@@ -19,70 +19,5 @@ app.use("/", indexRouter);
 app.use("/users", usersRouter);
 
 app.use("/api", apiRouter);
-
-apiRouter.post("/second", function (req, res) {
-  const responseBody = {
-    version: "2.0",
-    template: {
-      outputs: [
-        {
-          simpleText: {
-            text: "두번째 블록으로 왔숨다",
-          },
-        },
-      ],
-      quickReplies: [
-        {
-          messageText: "adela",
-          action: "message",
-          label: "adela",
-        },
-      ],
-    },
-  };
-});
-
-apiRouter.post("/", function (req, res) {
-  console.log(req.body);
-  const responseBody = {
-    version: "2.0",
-    template: {
-      outputs: [
-        {
-          simpleText: {
-            text: "토트넘 선수 리스트입니다.",
-          },
-        },
-      ],
-      quickReplies: [
-        {
-          messageText: "손흥민",
-          action: "block",
-          blockId: "5e1728ebffa7480001c29ffb",
-          label: "손흥민",
-        },
-        {
-          messageText: "헤리케인",
-          action: "block",
-          blockId: "5e1728ebffa7480001c29ffb",
-          label: "헤리케인",
-        },
-        {
-          messageText: "에릭센",
-          action: "block",
-          blockId: "5e1728ebffa7480001c29ffb",
-          label: "에릭센",
-        },
-        {
-          messageText: "요리스",
-          action: "block",
-          blockId: "5e1728ebffa7480001c29ffb",
-          label: "요리스",
-        },
-      ],
-    },
-  };
-  res.status(200).send(responseBody);
-});
 
 module.exports = app;
